完善代码，修复bug

+++ manager-service/src/main/java/com/jenschen/service/impl/RoleServiceImpl.java
@@ -3,9 +3,11 @@ package com.jenschen.service.impl;
 import cn.hutool.core.bean.BeanUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.jenschen.base.Response;
+import com.jenschen.constant.CommonConstant;
 import com.jenschen.entity.PermissionEntity;
 import com.jenschen.entity.RoleEntity;
 import com.jenschen.entity.RolePermissionEntity;
+import com.jenschen.enumeration.ErrorEnum;
 import com.jenschen.mapper.RoleMapper;
 import com.jenschen.mapper.RolePermissionMapper;
 import com.jenschen.mapper.RoleUserMapper;
@@ -63,6 +65,11 @@ public class RoleServiceImpl extends AbstractService<RoleEntity> implements Role
 
     @Override
     public Response<Object> edit(RoleReq roleReq) {
+        //默认角色admin禁止编辑
+        if(roleReq.getId() == CommonConstant.DEFAULT_ROLE_ADMIN_ID){
+            return ResultUtil.error(ErrorEnum.ACCESS_DENIED);
+        }
+
         RoleEntity role = BeanUtil.copyProperties(roleReq, RoleEntity.class);
         role.updated(LocalDateTime.now(), 1);
         roleMapper.updateById(role);
@@ -72,6 +79,11 @@ public class RoleServiceImpl extends AbstractService<RoleEntity> implements Role
 
     @Override
     public Response<Object> delete(Integer id) {
+        //默认角色admin禁止删除
+        if(id == CommonConstant.DEFAULT_ROLE_ADMIN_ID){
+            return ResultUtil.error(ErrorEnum.ACCESS_DENIED);
+        }
+
         RoleEntity entity = RoleEntity.builder().build();
         entity.setId(id);
         entity.deleted(LocalDateTime.now(), 1);
@@ -85,6 +97,11 @@ public class RoleServiceImpl extends AbstractService<RoleEntity> implements Role
 
     @Override
     public Response<Object> updatePermissions(RolePermissionReq rolePermissionReq) {
+        //默认角色admin禁止删除
+        if(rolePermissionReq.getRoleId() == CommonConstant.DEFAULT_ROLE_ADMIN_ID){
+            return ResultUtil.error(ErrorEnum.ACCESS_DENIED);
+        }
+
         //删除角色权限表中的数据
         rolePermissionMapper.deletePermissionByRoleId(rolePermissionReq.getRoleId(), 1);
